Factor out utilities for schemas
